Api

Documentation for Oxygen.jl

Starting the webserver

Oxygen.Core.serve — Function

serve(; middleware::Vector=[], host="127.0.0.1", port=8080, serialize=true, async=false, kwargs...)

Start the webserver with your own custom request handler

Oxygen.Core.serveparallel — Function

serveparallel(; middleware::Vector=[], host="127.0.0.1", port=8080, queuesize=1024, serialize=true, async=false, kwargs...)

Starts the webserver in streaming mode with your own custom request handler and spawns n - 1 worker threads to process individual requests. A Channel is used to schedule individual requests in FIFO order. Requests in the channel are then removed & handled by each the worker threads asynchronously.

Routing

Oxygen.Core.@get — Macro

@get(path::String, func::Function)

Used to register a function to a specific endpoint to handle GET requests

Oxygen.Core.@post — Macro

@post(path::String, func::Function)

Used to register a function to a specific endpoint to handle POST requests

Oxygen.Core.@put — Macro

@put(path::String, func::Function)

Used to register a function to a specific endpoint to handle PUT requests

Oxygen.Core.@patch — Macro

@patch(path::String, func::Function)

Used to register a function to a specific endpoint to handle PATCH requests

Oxygen.Core.@delete — Macro

@delete(path::String, func::Function)

Used to register a function to a specific endpoint to handle DELETE requests

Oxygen.Core.@route — Macro

@route(methods::Array{String}, path::String, func::Function)

Used to register a function to a specific endpoint to handle mulitiple request types

Mounting Files

Oxygen.Core.@staticfiles — Macro

@staticfiles(folder::String, mountdir::String)

Mount all files inside the /static folder (or user defined mount point)

Oxygen.Core.@dynamicfiles — Macro

@dynamicfiles(folder::String, mountdir::String)

Mount all files inside the /static folder (or user defined mount point), but files are re-read on each request

Oxygen.Core.FileUtil.file — Function

file(filepath::String)

Reads a file as a String

Autogenerated Docs

Oxygen.Core.AutoDoc.configdocs — Function

configdocs(docs_url::String = "/docs", schema_url::String = "/schema")

Configure the default docs and schema endpoints

Oxygen.Core.AutoDoc.enabledocs — Function

enabledocs()

Tells the api to mount the api doc endpoints on startup

Oxygen.Core.AutoDoc.disabledocs — Function

disabledocs()

Tells the api to SKIP mounting the api doc endpoints on startup

Oxygen.Core.AutoDoc.isdocsenabled — Function

isdocsenabled()

Returns true if we should mount the api doc endpoints, false otherwise

Oxygen.Core.AutoDoc.mergeschema — Function

mergeschema(route::String, customschema::Dict)

Merge the schema of a specific route

mergeschema(customschema::Dict)

Merge the top-level autogenerated schema with a custom schema

Oxygen.Core.AutoDoc.setschema — Function

setschema(customschema::Dict)

Overwrites the entire internal schema

Oxygen.Core.AutoDoc.getschema — Function

getschema()

Return the current internal schema for this app

Missing docstring.

Missing docstring for enabledocs. Check Documenter's build log for details.

Missing docstring.

Missing docstring for disabledocs. Check Documenter's build log for details.

Missing docstring.

Missing docstring for isdocsenabled. Check Documenter's build log for details.

Helper functions

Oxygen.Util.queryparams — Function

queryparams(request::HTTP.Request)

Parse's the query parameters from the Requests URL and return them as a Dict

Oxygen.Util.html — Function

html(content::String; status::Int, headers::Pair)

A convenience funtion to return a String that should be interpreted as HTML

Oxygen.BodyParsers.text — Function

text(request::HTTP.Request)

Read the body of a HTTP.Request as a String

text(response::HTTP.Response)

Read the body of a HTTP.Response as a String

Oxygen.BodyParsers.json — Function

json(request::HTTP.Request)

Read the body of a HTTP.Request as JSON

json(request::HTTP.Request, classtype)

Read the body of a HTTP.Request as JSON and serialize it into a custom struct

json(response::HTTP.Response)

Read the body of a HTTP.Response as JSON

json(response::HTTP.Response, classtype)

Read the body of a HTTP.Response as JSON and serialize it into a custom struct

Oxygen.BodyParsers.binary — Function

binary(request::HTTP.Request)

Read the body of a HTTP.Request as a Vector{UInt8}

Extra's

Oxygen.Core.AutoDoc.router — Function

This functions assists registering routes with a specific prefix. You can optionally assign tags either at the prefix and/or route level which are used to group and organize the autogenerated documentation

Oxygen.Core.internalrequest — Function

internalrequest(req::HTTP.Request; middleware::Vector=[], serialize::Bool=true)

Directly call one of our other endpoints registered with the router, using your own middleware and bypassing any globally defined middleware

Oxygen.Util.redirect — Function

redirect(path::String; code = 308)

return a redirect response

Oxygen.Core.terminate — Function

terminate()

stops the webserver immediately

Oxygen.Core.starttasks — Function

starttasks()

Start all background repeat tasks

Oxygen.Core.stoptasks — Function

stoptasks()

Stop all background repeat tasks

Oxygen.Core.resetstate — Function

Reset all the internal state variables