MolOps

Operations to process and extract information from OxygenMols.

Oxygen.molfeatures — Method

molfeatures(smiles)

Creates atom-wise features matrix for a molecule based on atomtypes, valence, and hybridization. Note: molfeatures and getadjmatrix do not return padded arrays.

Examples

julia> molfeatures(smilestomol("c1ccccc1"))
6×28 Array{Float32,2}:
 1.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  1.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  1.0  0.0  0.0
 1.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  1.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  1.0  0.0  0.0
 1.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  1.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  1.0  0.0  0.0
 1.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  1.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  1.0  0.0  0.0
 1.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  1.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  1.0  0.0  0.0
 1.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  1.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  1.0  0.0  0.0

Oxygen.atomtypes — Method

atomtypes(smiles)

Creates one-hot mapping of atoms in a molecule to common atoms. Current valid atoms array is ["C", "N", "O", "S", "F", "P", "Cl", "Mg", "Na", "Br", "Ca", "Cu", "Pd", "I", "Al"].

Examples

julia> atomtypes(smilestomol("c1ccccc1"))
6×15 Array{Float32,2}:
 1.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0
 1.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0
 1.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0
 1.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0
 1.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0
 1.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0

Oxygen.getadjmatrix — Method

getadjmatrix(smiles)

Generates adjacency matrix of molecule, weighted by bond type

Examples

julia> getadjmatrix(smilestomol("c1ccccc1"))
6×6 Array{Float32,2}:
 0.0  1.5  0.0  0.0  0.0  1.5
 1.5  0.0  1.5  0.0  0.0  0.0
 0.0  1.5  0.0  1.5  0.0  0.0
 0.0  0.0  1.5  0.0  1.5  0.0
 0.0  0.0  0.0  1.5  0.0  1.5
 1.5  0.0  0.0  0.0  1.5  0.0

Oxygen.getbonds — Method

getbonds(mol)

Retrieves list of bonds in a molecule. Bond list formatted as (starting atom, ending atom, bond type). All bonds are unique.

Examples

julia> getbonds(smilestomol("c1ccccc1"))
6-element Array{Tuple,1}:
 (1, 2, 1.5)
 (1, 6, 1.5)
 (2, 3, 1.5)
 (3, 4, 1.5)
 (4, 5, 1.5)
 (5, 6, 1.5)

Oxygen.hybridization — Method

hybridization(smiles)

Creates one-hot mapping of atom hybridization states in a molecule. Currently creates onehot mapping of bonding electron count. Hybridization states will be added soon.

Examples

``` julia> hybridization(smilestomol("c1ccccc1")) 6×7 Array{Int64,2}: 0 0 0 0 1 0 0 0 0 0 0 1 0 0 0 0 0 0 1 0 0 0 0 0 0 1 0 0 0 0 0 0 1 0 0 0 0 0 0 1 0 0

Oxygen.valencies — Method

valencies(smiles)

Creates one-hot mapping of atom valencies in a molecule. Current maximum valid valencies (number of connected non-hydrogen atoms) is 6.

Examples

julia> valencies(smilestomol("c1ccccc1"))
6×6 Array{Int64,2}:
 0  1  0  0  0  0
 0  1  0  0  0  0
 0  1  0  0  0  0
 0  1  0  0  0  0
 0  1  0  0  0  0
 0  1  0  0  0  0