OxygenMol

Data types for organic molecules.

OxygenMol implements a mutable molecule type, using weighted adjacency lists to denote bonds. A bond order of 1.5 refers to an aromatic bond. Hydrogens are implicit.

Oxygen.OxygenMol — Type

OxygenMol

Mutable molecule type.

Represents molecules using an array of atoms and a weighted adjacency list to represent bonds.