Api

Documentation for Oxygen.jl

Starting the webserver

Oxygen.serve — Function

serve(host="127.0.0.1", port=8080; kwargs...)

Start the webserver with the default request handler

Oxygen.serve — Function

serve(handler::Function, host="127.0.0.1", port=8080; kwargs...)

Start the webserver with your own custom request handler

Oxygen.serveparallel — Function

serveparallel(host="127.0.0.1", port=8080, queuesize=1024; kwargs...)

Starts the webserver in streaming mode and spawns n - 1 worker threads to process individual requests. A Channel is used to schedule individual requests in FIFO order. Requests in the channel are then removed & handled by each the worker threads asynchronously.

Oxygen.serveparallel — Function

serveparallel(handler::Function, host="127.0.0.1", port=8080, queuesize=1024; kwargs...)

Starts the webserver in streaming mode with your own custom request handler and spawns n - 1 worker threads to process individual requests. A Channel is used to schedule individual requests in FIFO order. Requests in the channel are then removed & handled by each the worker threads asynchronously.

Routing

Oxygen.@get — Macro

@get(path::String, func::Function)

Used to register a function to a specific endpoint to handle GET requests

Oxygen.@post — Macro

@post(path::String, func::Function)

Used to register a function to a specific endpoint to handle POST requests

Oxygen.@put — Macro

@put(path::String, func::Function)

Used to register a function to a specific endpoint to handle PUT requests

Oxygen.@patch — Macro

@patch(path::String, func::Function)

Used to register a function to a specific endpoint to handle PATCH requests

Oxygen.@delete — Macro

@delete(path::String, func::Function)

Used to register a function to a specific endpoint to handle DELETE requests

Oxygen.@route — Macro

@route(methods::Array{String}, path::String, func::Function)

Used to register a function to a specific endpoint to handle mulitiple request types

Oxygen.@register — Macro

@register(httpmethod::String, path::String, func::Function)

Register a request handler function with a path to the ROUTER

Mounting Files

Oxygen.@staticfiles — Macro

@staticfiles(folder::String, mountdir::String)

Mount all files inside the /static folder (or user defined mount point)

Oxygen.@dynamicfiles — Macro

@dynamicfiles(folder::String, mountdir::String)

Mount all files inside the /static folder (or user defined mount point), but files are re-read on each request

Oxygen.FileUtil.file — Function

file(folder::String, filename::String)

Read in a file as a String (from the projects current working directory)

file(folder::String, filename::String)

Read in a file from a specific folder as a String (from the projects current working directory)

Helper functions

Oxygen.queryparams — Function

queryparams(request::HTTP.Request)

Parse's the query parameters from the Requests URL and return them as a Dict

Oxygen.html — Function

html(content::String; status::Int, headers::Pair)

A convenience funtion to return a String that should be interpreted as HTML

Oxygen.text — Function

text(request::HTTP.Request)

Read the body of a HTTP.Request as a String

text(response::HTTP.Response)

Read the body of a HTTP.Response as a String

Oxygen.json — Function

json(request::HTTP.Request)

Read the body of a HTTP.Request as JSON

json(request::HTTP.Request, classtype)

Read the body of a HTTP.Request as JSON and serialize it into a custom struct

json(response::HTTP.Response)

Read the body of a HTTP.Response as JSON

json(response::HTTP.Response, classtype)

Read the body of a HTTP.Response as JSON and serialize it into a custom struct

Oxygen.binary — Function

binary(request::HTTP.Request)

Read the body of a HTTP.Request as a Vector{UInt8}

Extra's

Oxygen.internalrequest — Function

internalrequest(request::HTTP.Request)

Directly call one of our other endpoints registered with the router

Oxygen.terminate — Method

terminate()

stops the webserver immediately