Api

Documentation for Oxygen.jl

Starting the webserver

Oxygen.ServerUtil.serve — Function

serve(; host="127.0.0.1", port=8080, kwargs...)

Start the webserver with the default request handler

serve(handler::Function; host="127.0.0.1", port=8080, kwargs...)

Start the webserver with your own custom request handler

Oxygen.ServerUtil.serveparallel — Function

serveparallel(; host="127.0.0.1", port=8080, queuesize=1024, kwargs...)

Starts the webserver in streaming mode and spawns n - 1 worker threads to process individual requests. A Channel is used to schedule individual requests in FIFO order. Requests in the channel are then removed & handled by each the worker threads asynchronously.

serveparallel(handler::Function; host="127.0.0.1", port=8080, queuesize=1024, kwargs...)

Starts the webserver in streaming mode with your own custom request handler and spawns n - 1 worker threads to process individual requests. A Channel is used to schedule individual requests in FIFO order. Requests in the channel are then removed & handled by each the worker threads asynchronously.

Routing

Oxygen.ServerUtil.@get — Macro

@get(path::String, func::Function)

Used to register a function to a specific endpoint to handle GET requests

Oxygen.ServerUtil.@post — Macro

@post(path::String, func::Function)

Used to register a function to a specific endpoint to handle POST requests

Oxygen.ServerUtil.@put — Macro

@put(path::String, func::Function)

Used to register a function to a specific endpoint to handle PUT requests

Oxygen.ServerUtil.@patch — Macro

@patch(path::String, func::Function)

Used to register a function to a specific endpoint to handle PATCH requests

Oxygen.ServerUtil.@delete — Macro

@delete(path::String, func::Function)

Used to register a function to a specific endpoint to handle DELETE requests

Oxygen.ServerUtil.@route — Macro

@route(methods::Array{String}, path::String, func::Function)

Used to register a function to a specific endpoint to handle mulitiple request types

Mounting Files

Oxygen.ServerUtil.@staticfiles — Macro

@staticfiles(folder::String, mountdir::String)

Mount all files inside the /static folder (or user defined mount point)

Oxygen.ServerUtil.@dynamicfiles — Macro

@dynamicfiles(folder::String, mountdir::String)

Mount all files inside the /static folder (or user defined mount point), but files are re-read on each request

Oxygen.FileUtil.file — Function

file(filepath::String)

Reads a file as a String

Swagger Docs

Oxygen.ServerUtil.AutoDoc.configdocs — Function

configdocs(docs_url::String = "/docs", schema_url::String = "/docs/schema")

Configure the default docs and schema endpoints

Oxygen.ServerUtil.AutoDoc.enabledocs — Function

enabledocs()

Tells the api to mount the api doc endpoints on startup

Oxygen.ServerUtil.AutoDoc.disabledocs — Function

disabledocs()

Tells the api to SKIP mounting the api doc endpoints on startup

Oxygen.ServerUtil.AutoDoc.isdocsenabled — Function

isdocsenabled()

Returns true if we should mount the api doc endpoints, false otherwise

Oxygen.ServerUtil.AutoDoc.mergeschema — Function

mergeschema(route::String, customschema::Dict)

Merge the schema of a specific route

mergeschema(customschema::Dict)

Merge the top-level autogenerated schema with a custom schema

Oxygen.ServerUtil.AutoDoc.setschema — Function

setschema(customschema::Dict)

Overwrites the entire internal schema

Oxygen.ServerUtil.AutoDoc.getschema — Function

getschema()

Return the current internal schema for this app

Helper functions

Oxygen.Util.queryparams — Function

queryparams(request::HTTP.Request)

Parse's the query parameters from the Requests URL and return them as a Dict

Oxygen.Util.html — Function

html(content::String; status::Int, headers::Pair)

A convenience funtion to return a String that should be interpreted as HTML

Oxygen.BodyParsers.text — Function

text(request::HTTP.Request)

Read the body of a HTTP.Request as a String

text(response::HTTP.Response)

Read the body of a HTTP.Response as a String

Oxygen.BodyParsers.json — Function

json(request::HTTP.Request)

Read the body of a HTTP.Request as JSON

json(request::HTTP.Request, classtype)

Read the body of a HTTP.Request as JSON and serialize it into a custom struct

json(response::HTTP.Response)

Read the body of a HTTP.Response as JSON

json(response::HTTP.Response, classtype)

Read the body of a HTTP.Response as JSON and serialize it into a custom struct

Oxygen.BodyParsers.binary — Function

binary(request::HTTP.Request)

Read the body of a HTTP.Request as a Vector{UInt8}

Extra's

Oxygen.ServerUtil.internalrequest — Function

internalrequest(request::HTTP.Request)

Directly call one of our other endpoints registered with the router

internalrequest(request::HTTP.Request, handler::Function)

Directly call one of our other endpoints registered with the router, using your own Handler function

Oxygen.ServerUtil.terminate — Method

terminate()

stops the webserver immediately