Api

Documentation for Oxygen.jl

Starting the webserver

Oxygen.serve — Function

serve(; middleware::Vector=[], handler=stream_handler, host="127.0.0.1", port=8080, async=false, parallel=false, serialize=true, catch_errors=true, docs=true, metrics=true, show_errors=true, show_banner=true, docs_path="/docs", schema_path="/schema", external_url=nothing, revise, kwargs...)

Start the webserver with your own custom request handler

Oxygen.serveparallel — Function

serveparallel(; middleware::Vector=[], handler=stream_handler, host="127.0.0.1", port=8080, serialize=true, async=false, catch_errors=true, docs=true, metrics=true, kwargs...)

Routing

Oxygen.@get — Macro

@get(path::String, func::Function)

Used to register a function to a specific endpoint to handle GET requests

Oxygen.@post — Macro

@post(path::String, func::Function)

Used to register a function to a specific endpoint to handle POST requests

Oxygen.@put — Macro

@put(path::String, func::Function)

Used to register a function to a specific endpoint to handle PUT requests

Oxygen.@patch — Macro

@patch(path::String, func::Function)

Used to register a function to a specific endpoint to handle PATCH requests

Oxygen.@delete — Macro

@delete(path::String, func::Function)

Used to register a function to a specific endpoint to handle DELETE requests

Oxygen.@route — Macro

@route(methods::Array{String}, path::String, func::Function)

Used to register a function to a specific endpoint to handle mulitiple request types

Missing docstring.

Missing docstring for get(path, func). Check Documenter's build log for details.

Missing docstring.

Missing docstring for post(path, func). Check Documenter's build log for details.

Missing docstring.

Missing docstring for put(path, func). Check Documenter's build log for details.

Missing docstring.

Missing docstring for patch(path, func). Check Documenter's build log for details.

Missing docstring.

Missing docstring for delete(path, func). Check Documenter's build log for details.

Missing docstring.

Missing docstring for route(methods, path, func). Check Documenter's build log for details.

Mounting Files

Oxygen.@staticfiles — Macro

@staticfiles(folder::String, mountdir::String, headers::Vector{Pair{String,String}}=[])

Mount all files inside the /static folder (or user defined mount point)

Oxygen.@dynamicfiles — Macro

@dynamicfiles(folder::String, mountdir::String, headers::Vector{Pair{String,String}}=[])

Mount all files inside the /static folder (or user defined mount point), but files are re-read on each request

Oxygen.staticfiles — Function

staticfiles(folder::String, mountdir::String; headers::Vector{Pair{String,String}}=[], loadfile::Union{Function,Nothing}=nothing)

Mount all files inside the /static folder (or user defined mount point). The headers array will get applied to all mounted files

Oxygen.dynamicfiles — Function

dynamicfiles(folder::String, mountdir::String; headers::Vector{Pair{String,String}}=[], loadfile::Union{Function,Nothing}=nothing)

Mount all files inside the /static folder (or user defined mount point), but files are re-read on each request. The headers array will get applied to all mounted files

Autogenerated Docs

Oxygen.configdocs — Function

configdocs(docspath::String = "/docs", schemapath::String = "/schema")

Configure the default docs and schema endpoints

Missing docstring.

Missing docstring for enabledocs. Check Documenter's build log for details.

Missing docstring.

Missing docstring for disabledocs. Check Documenter's build log for details.

Missing docstring.

Missing docstring for isdocsenabled. Check Documenter's build log for details.

Oxygen.mergeschema — Function

mergeschema(route::String, customschema::Dict)

Merge the schema of a specific route

Oxygen.setschema — Function

setschema(customschema::Dict)

Overwrites the entire internal schema

Oxygen.getschema — Function

getschema()

Return the current internal schema for this app

Helper functions

Missing docstring.

Missing docstring for queryparams. Check Documenter's build log for details.

Oxygen.Core.Util.formdata — Function

formdata(request::HTTP.Request)

Read the html form data from the body of a HTTP.Request

formdata(request::HTTP.Response)

Read the html form data from the body of a HTTP.Response

Oxygen.Core.Util.html — Function

html(content::String; status::Int, headers::Vector{Pair}) :: HTTP.Response

A convenience function to return a String that should be interpreted as HTML

Oxygen.Core.Util.text — Function

text(request::HTTP.Request)

Read the body of a HTTP.Request as a String

text(response::HTTP.Response)

Read the body of a HTTP.Response as a String

text(content::String; status::Int, headers::Vector{Pair}) :: HTTP.Response

A convenience function to return a String that should be interpreted as plain text

Oxygen.Core.Util.file — Function

file(filepath::String; loadfile=nothing, status = 200, headers = []) :: HTTP.Response

Reads a file and returns a HTTP.Response. The file is read as binary. If the file does not exist, an ArgumentError is thrown. The MIME type and the size of the file are added to the headers.

Arguments

filepath: The path to the file to be read.

loadfile: An optional function to load the file. If not provided, the file is read using the open function.

status: The HTTP status code to be used in the response. Defaults to 200.

headers: Any additional headers to be included in the response. Defaults to an empty array.

Returns

A HTTP response.

Oxygen.Core.Util.xml — Function

xml(content::String; status::Int, headers::Vector{Pair}) :: HTTP.Response

A convenience function to return a String that should be interpreted as XML

Oxygen.Core.Util.js — Function

js(content::String; status::Int, headers::Vector{Pair}) :: HTTP.Response

A convenience function to return a String that should be interpreted as JavaScript

Oxygen.Core.Util.json — Function

json(request::HTTP.Request; keyword_arguments...)

Read the body of a HTTP.Request as JSON with additional arguments for the read/serializer.

json(request::HTTP.Request, class_type; keyword_arguments...)

Read the body of a HTTP.Request as JSON with additional arguments for the read/serializer into a custom struct.

json(response::HTTP.Response; keyword_arguments)

Read the body of a HTTP.Response as JSON with additional keyword arguments

json(response::HTTP.Response, class_type; keyword_arguments)

Read the body of a HTTP.Response as JSON with additional keyword arguments and serialize it into a custom struct

json(content::Any; status::Int, headers::Vector{Pair}) :: HTTP.Response

A convenience function to return a String that should be interpreted as JSON

json(content::Vector{UInt8}; status::Int, headers::Vector{Pair}) :: HTTP.Response

A helper function that can be passed binary data that should be interpreted as JSON. No conversion is done on the content since it's already in binary format.

Oxygen.Core.Util.css — Function

css(content::String; status::Int, headers::Vector{Pair}) :: HTTP.Response

A convenience function to return a String that should be interpreted as CSS

Oxygen.Core.Util.binary — Function

binary(request::HTTP.Request)

Read the body of a HTTP.Request as a Vector{UInt8}

binary(content::Vector{UInt8}; status::Int, headers::Vector{Pair}) :: HTTP.Response

A convenience function to return a Vector of UInt8 that should be interpreted as binary data

Repeat Tasks & Cron Scheduling

Oxygen.@cron — Macro

@cron(expression::String, func::Function)

Registers a function with a cron expression. This will extract either the function name or the random Id julia assigns to each lambda function.

@cron(expression::String, name::String, func::Function)

This variation provides way manually "name" a registered function. This information is used by the server on startup to log out all cron jobs.

Oxygen.starttasks — Function

starttasks()

Start all background repeat tasks

Oxygen.stoptasks — Function

stoptasks()

Stop all background repeat tasks

Oxygen.cleartasks — Function

cleartasks(ct::ServerContext)

Clear any stored repeat task definitions

Oxygen.startcronjobs — Function

startcronjobs()

Start all the cron cronjobs within their own async task. Each individual task will loop conintually and sleep untill the next time it's suppost to

Oxygen.stopcronjobs — Function

stopcronjobs()

Stop each background task by toggling a global reference that all cron jobs reference

Oxygen.clearcronjobs — Function

Clears all cron job defintions

Extra's

Oxygen.router — Function

Oxygen.internalrequest — Function

internalrequest(req::HTTP.Request; middleware::Vector=[], serialize::Bool=true, catch_errors::Bool=true)

Directly call one of our other endpoints registered with the router, using your own middleware and bypassing any globally defined middleware

Oxygen.Core.Util.redirect — Function

redirect(path::String; code = 307)

return a redirect response

Oxygen.terminate — Function

terminate(ctx)

Gracefully shuts down the webserver

Oxygen.resetstate — Function

resetstate()

Reset all the internal state variables