Api

Documentation for Oxygen.jl

Starting the webserver

Oxygen.Core.serve — Function

serve(; middleware::Vector=[], handler=stream_handler, host="127.0.0.1", port=8080, serialize=true, async=false, catch_errors=true, metrics=true, kwargs...)

Start the webserver with your own custom request handler

Oxygen.Core.serveparallel — Function

serveparallel(; middleware::Vector=[], handler=stream_handler, host="127.0.0.1", port=8080, queuesize=1024, serialize=true, async=false, catch_errors=true, metrics=true, kwargs...)

Starts the webserver in streaming mode with your own custom request handler and spawns n - 1 worker threads to process individual requests. A Channel is used to schedule individual requests in FIFO order. Requests in the channel are then removed & handled by each the worker threads asynchronously.

Routing

Oxygen.Core.@get — Macro

@get(path::String, func::Function)

Used to register a function to a specific endpoint to handle GET requests

Oxygen.Core.@post — Macro

@post(path::String, func::Function)

Used to register a function to a specific endpoint to handle POST requests

Oxygen.Core.@put — Macro

@put(path::String, func::Function)

Used to register a function to a specific endpoint to handle PUT requests

Oxygen.Core.@patch — Macro

@patch(path::String, func::Function)

Used to register a function to a specific endpoint to handle PATCH requests

Oxygen.Core.@delete — Macro

@delete(path::String, func::Function)

Used to register a function to a specific endpoint to handle DELETE requests

Oxygen.Core.@route — Macro

@route(methods::Array{String}, path::String, func::Function)

Used to register a function to a specific endpoint to handle mulitiple request types

Missing docstring.

Missing docstring for get(path, func). Check Documenter's build log for details.

Missing docstring.

Missing docstring for post(path, func). Check Documenter's build log for details.

Missing docstring.

Missing docstring for put(path, func). Check Documenter's build log for details.

Missing docstring.

Missing docstring for patch(path, func). Check Documenter's build log for details.

Missing docstring.

Missing docstring for delete(path, func). Check Documenter's build log for details.

Missing docstring.

Missing docstring for route(methods, path, func). Check Documenter's build log for details.

Mounting Files

Oxygen.Core.@staticfiles — Macro

@staticfiles(folder::String, mountdir::String, set_headers::Union{Function,Nothing}=nothing)

Mount all files inside the /static folder (or user defined mount point)

Oxygen.Core.@dynamicfiles — Macro

@dynamicfiles(folder::String, mountdir::String, set_headers::Union{Function,Nothing}=nothing)

Mount all files inside the /static folder (or user defined mount point), but files are re-read on each request

Oxygen.Core.staticfiles — Function

staticfiles(folder::String, mountdir::String; set_headers::Union{Function,Nothing}=nothing, loadfile::Union{Function,Nothing}=nothing)

Mount all files inside the /static folder (or user defined mount point). The set_headers callback is passed the route, mime type, and the current headers for the given resource and should return a finalized vector of pairs to use as the headers in the outgoing Request

Oxygen.Core.dynamicfiles — Function

dynamicfiles(folder::String, mountdir::String; set_headers::Union{Function,Nothing}=nothing, loadfile::Union{Function,Nothing}=nothing)

Mount all files inside the /static folder (or user defined mount point), but files are re-read on each request. The set_headers callback is passed the route, mime type, and the current headers for the given resource and should return a finalized vector of pairs to use as the headers in the outgoing Request.

Oxygen.Core.FileUtil.file — Function

file(filepath::String)

Reads a file as a String

Autogenerated Docs

Oxygen.Core.AutoDoc.configdocs — Function

configdocs(docs_url::String = "/docs", schema_url::String = "/schema")

Configure the default docs and schema endpoints

Oxygen.Core.AutoDoc.enabledocs — Function

enabledocs()

Tells the api to mount the api doc endpoints on startup

Oxygen.Core.AutoDoc.disabledocs — Function

disabledocs()

Tells the api to SKIP mounting the api doc endpoints on startup

Oxygen.Core.AutoDoc.isdocsenabled — Function

isdocsenabled()

Returns true if we should mount the api doc endpoints, false otherwise

Oxygen.Core.AutoDoc.mergeschema — Function

mergeschema(route::String, customschema::Dict)

Merge the schema of a specific route

mergeschema(customschema::Dict)

Merge the top-level autogenerated schema with a custom schema

Oxygen.Core.AutoDoc.setschema — Function

setschema(customschema::Dict)

Overwrites the entire internal schema

Oxygen.Core.AutoDoc.getschema — Function

getschema()

Return the current internal schema for this app

Missing docstring.

Missing docstring for enabledocs. Check Documenter's build log for details.

Missing docstring.

Missing docstring for disabledocs. Check Documenter's build log for details.

Missing docstring.

Missing docstring for isdocsenabled. Check Documenter's build log for details.

Helper functions

Oxygen.Core.Util.queryparams — Function

queryparams(request::HTTP.Request)

Parse's the query parameters from the Requests URL and return them as a Dict

Oxygen.Core.Util.html — Function

html(content::String; status::Int, headers::Pair)

A convenience funtion to return a String that should be interpreted as HTML

Oxygen.Core.BodyParsers.text — Function

text(request::HTTP.Request)

Read the body of a HTTP.Request as a String

text(response::HTTP.Response)

Read the body of a HTTP.Response as a String

Oxygen.Core.BodyParsers.json — Function

json(request::HTTP.Request; keyword_arguments...)

Read the body of a HTTP.Request as JSON with additional arguments for the read/serializer.

json(request::HTTP.Request, classtype; keyword_arguments...)

Read the body of a HTTP.Request as JSON with additional arguments for the read/serializer into a custom struct.

json(response::HTTP.Response; keyword_arguments)

Read the body of a HTTP.Response as JSON with additional keyword arguments

json(response::HTTP.Response, classtype; keyword_arguments)

Read the body of a HTTP.Response as JSON with additional keyword arguments and serialize it into a custom struct

Oxygen.Core.BodyParsers.binary — Function

binary(request::HTTP.Request)

Read the body of a HTTP.Request as a Vector{UInt8}

Repeat Tasks & Cron Scheduling

Oxygen.Core.AutoDoc.Cron.@cron — Macro

@cron(expression::String, func::Function)

Registers a function with a cron expression. This will extract either the function name or the random Id julia assigns to each lambda function.

@cron(expression::String, name::String, func::Function)

This variation Provide another way manually "name" a registered function. This information is used by the server on startup to log out all cron jobs.

Oxygen.Core.starttasks — Function

starttasks()

Start all background repeat tasks

Oxygen.Core.stoptasks — Function

stoptasks()

Stop all background repeat tasks

Oxygen.Core.AutoDoc.Cron.startcronjobs — Function

startcronjobs()

Start all the cron job_definitions within their own async task. Each individual task will loop conintually and sleep untill the next time it's suppost to

Oxygen.Core.AutoDoc.Cron.stopcronjobs — Function

stopcronjobs()

Stop each background task by toggling a global reference that all cron jobs reference

Extra's

Oxygen.Core.AutoDoc.router — Function

This functions assists registering routes with a specific prefix. You can optionally assign tags either at the prefix and/or route level which are used to group and organize the autogenerated documentation

Oxygen.Core.internalrequest — Function

internalrequest(req::HTTP.Request; middleware::Vector=[], serialize::Bool=true, catch_errors::Bool=true)

Directly call one of our other endpoints registered with the router, using your own middleware and bypassing any globally defined middleware

Oxygen.Core.Util.redirect — Function

redirect(path::String; code = 308)

return a redirect response

Oxygen.Core.terminate — Function

terminate()

stops the webserver immediately

Oxygen.Core.resetstate — Function

Reset all the internal state variables